requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
idna==2.8
multidict==4.7.5
requests==2.22.0
six==1.14.0
slackclient==2.5.0
spotipy==2.7.1
urllib3==1.25.8
yarl==1.4.2
